requirements.txt
# requirements.txt
fastmcp>=0.1.0
ollama>=0.1.0
pydantic>=2.0.0
aiohttp>=3.8.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# requirements.txt - Complete dependencies
fastmcp>=0.1.0
ollama>=0.1.0
pydantic>=2.0.0
aiohttp>=3.8.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
websockets>=12.0
plotly>=5.18.0
prometheus-client>=0.17.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=5.0.0
celery>=5.3.0
click>=8.1.0
colorlog>=6.7.0
pyyaml>=6.0
ujson>=5.7.0
orjson>=3.9.0
cryptography>=41.0.0
jwt>=1.3.0
python-telegram-bot>=20.0
slack-sdk>=3.21.0
discord-webhook>=1.3.0